+++ packages/dara-components/changelog.md
@@ -6,6 +6,7 @@ title: Changelog
 
 -   Updated `Code` component to support copy to clipboard functionality.
 -   Updated `Code` component default themes.
+-   Updated `Code` component to display current language.
 
 ## 1.10.6